cleanup keystore files
